Moved S3 utils
