+++ scripts/extract_block.py
@@ -1,14 +1,9 @@
 import os
-import sys
 import argparse
 import numpy as np
-import z5py
 import zarr
 
-from inspect import getsourcefile
-
-sys.path.append(os.path.join(os.path.dirname(getsourcefile(lambda:0)), "prediction"))
-import upload_to_s3
+import flamingo_tools.s3_utils as s3_utils
 
 """
 This script extracts data around an input center coordinate in a given ROI halo.
@@ -71,9 +66,9 @@ def main(
     roi = tuple(slice(co - rh, co + rh) for co, rh in zip(coords, roi_halo))
 
     if s3:
-        bucket_name, service_endpoint, credentials = upload_to_s3.check_s3_credentials(s3_bucket_name, s3_service_endpoint, s3_credentials)
+        bucket_name, service_endpoint, credentials = s3_utils.check_s3_credentials(s3_bucket_name, s3_service_endpoint, s3_credentials)
 
-        s3_path, fs = upload_to_s3.get_s3_path(input_file, bucket_name=bucket_name, service_endpoint=service_endpoint, credential_file=credentials)
+        s3_path, fs = s3_utils.get_s3_path(input_file, bucket_name=bucket_name, service_endpoint=service_endpoint, credential_file=credentials)
 
         with zarr.open(s3_path, mode="r") as f:
             raw = f[input_key][roi]
