+++ scripts/prediction/postprocess_seg.py
@@ -5,10 +5,9 @@ import sys
 import pandas as pd
 import zarr
 
-import upload_to_s3
-
 sys.path.append("../..")
 
+import flamingo_tools.s3_utils as s3_utils
 
 def main():
     from flamingo_tools.segmentation import filter_isolated_objects
@@ -38,14 +37,14 @@ def main():
     tsv_table=None
 
     if args.s3_input is not None:
-        bucket_name, service_endpoint, credentials = upload_to_s3.check_s3_credentials(args.s3_bucket_name, args.s3_service_endpoint, args.s3_credentials)
+        bucket_name, service_endpoint, credentials = s3_utils.check_s3_credentials(args.s3_bucket_name, args.s3_service_endpoint, args.s3_credentials)
 
-        s3_path, fs = upload_to_s3.get_s3_path(args.s3_input, bucket_name=bucket_name, service_endpoint=service_endpoint, credential_file=credentials)
+        s3_path, fs = s3_utils.get_s3_path(args.s3_input, bucket_name=bucket_name, service_endpoint=service_endpoint, credential_file=credentials)
         with zarr.open(s3_path, mode="r") as f:
             segmentation = f[args.input_key]
 
         if args.tsv is not None:
-            tsv_path, fs = upload_to_s3.get_s3_path(args.tsv, bucket_name=bucket_name, service_endpoint=service_endpoint, credential_file=credentials)
+            tsv_path, fs = s3_utils.get_s3_path(args.tsv, bucket_name=bucket_name, service_endpoint=service_endpoint, credential_file=credentials)
             with fs.open(tsv_path, 'r') as f:
                 tsv_table = pd.read_csv(f, sep="\t")
